Update encrypted images
This change adds the update-encrypted-ami option to the brkt_cli.
It allows a user to retrieve a metavisor upgrader AMI and use it
to upgrade an already encrypted AMI. This provides a speed
improvement for users who wish to update their encrypted AMIs.

The user must specify the metavisor updater AMI as an argument.
This requirement will be taken away in a near-future commit.