Use 'sd' names for HVM instances. More validation of guest AMIs
The AWS GUI for HVM instances requires the use of 'sd' names as
attach points even though these names are encoded in Xen and
end up as 'xvd' names on the guest/NetBSD. If we use 'sd' names
we can avoid the common error case of the user picking 'sdf'
as an attach point even though this is being used by the metavisor
as 'xvdf'

Also handle a case where the root_device_name is not found
in the block_device_mapping for an AMI. We can't encrypt this
AMI in that form. Typically the root device name is '/dev/sda1'
and you'll find that in the block device map. If that's not the
case we can't rip out old volumes and replace them with new
volumes due to a bug in AWS (when you remove all the drives
from a stopped instance it still believes the root_device_name
is in use)

Finally launch m3.medium instances for the guest instances. They
boot faster and are cheaper than c3.xlarge instances