Log tracebacks at debug level
In some cases, we log tracebacks only when verbose logging is enabled.
We used to log the exception if log.isEnabledFor(logging.DEBUG) returned
True.

This code broke when we started writing debug log output to a temp file
(commit 9b4a6538d256602d8d5b885441858e6e8f6bc4d6).  After the logging
change, the call to log.isEnabledFor(logging.DEBUG) always returned
True.

With this commit, we fix all of the callsites, so that they log the
traceback with log.debug(..., exc_info=1).  This has the desired effect
of always writing tracebacks to the temp file, while only writing it to
stderr when the verbose logging flag is set.