Add HTTP(s) proxy for S3 connections
A customer wants to use a proxy to connect to the internet
from the host which is running the brkt-cli, while all other
(internal) connections do not require the proxy. The outside
(internet) connection is to S3 while the internal connections
goes to the local vCenter (which at most times has a well known
host name), to an ESX host on which the encryptor/updater VM
is deployed (whose IP most times is not known, but the list
of ESX hosts may available)  and to the encryptor VM itself
(whose IP comes from DHCP, hence is definitely not known
beforehand). This change allows the user to specify a HTTP(s)
proxy and port to allow the brkt-cli to use it for the
outbound connection to S3.