Don't log user-data unless --verbose was specified
Check the state of the --verbose option instead of the log level before
writing user-data to a temp file.

With commit 9b4a6538d256602d8d5b885441858e6e8f6bc4d6, we started writing
the debug log to a temp file.  This caused
log.isEnabledFor(logging.DEBUG) to always return true.  The side effect
is that we started always writing user-data to a temp file.