Delete boot disk by default for GCE instances
By default, we were not deleting the boot disk for a GCE instance
launched using the "brkt gce launch" command, when the instance was
terminated. This change makes the boot disk deletion the default
behaviour and toggles the purpose of the --delete-boot argument.
Now if this argument is specified, then we will retain the boot
disk on instance termination.