Add HTTP(s) proxy for S3 connections
This change allows the user to specify a HTTP(s) proxy and port to allow the brkt-cli to
use it for outbound connections to S3 when encrypting or updating images directly against
an ESX host. This is to bring it on par to when performing similar operations using a
vCenter host.

A customer wants to use a proxy to connect to the internet from the host which is running
the brkt-cli, while all other (internal) connections do not require the proxy. The outside
(internet) connection is to S3 while the internal connections goes to the local vCenter or
ESX host (which at most times has a well known host name).